feat: appNav와 mainactivity 파일 분리 및 최대한 자연스러운 화면 이동 구현
